refactoring(redis): move redis config to own configuration
